reduce js file number
