+++ myindex.ejs
@@ -2,7 +2,7 @@
   <html lang="en" dir="ltr">
     <head>
       <meta charset="utf-8">
-      <title>my paper patatap</title>
+      <title>Color Tap</title>
       <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css" integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
   <link rel="stylesheet" href="./styles/main.css">
     </head>
